Fix oversight in Linux build commands.
CXXFLAGS is LL_BUILD_RELEASE plus -m64; CFLAGS is supposed to be all of that
minus the -std=c++NN switch. Instead, CFLAGS was specified in terms of (unset)
CFLAGS.

Also, instead of hard-coding msbuild PlatformToolset as v143, use
AUTOBUILD_WIN_VSTOOLSET for compatibility with future updates to the VS
toolchain.